Add test case for browserify builds
This runs test cases:

 - on the source code (as was done before the change)
 - on the dist files (since these are what is actually used by users)
 - on the dist files after re-browserifying (this currently fails)
