+++ package.json
@@ -25,8 +25,10 @@
   },
   "license": "MIT",
   "scripts": {
-    "prepublish": "node bin/build-acorn.js && node bin/without_eval > dist/acorn_csp.js",
-    "test": "node test/run.js"
+    "build": "node bin/build-acorn.js && node bin/without_eval > dist/acorn_csp.js",
+    "prepublish": "npm run build",
+    "pretest": "npm run build && browserify test/run-dist.js > test/browserify.js",
+    "test": "node test/run.js && node test/run-dist.js && node test/browserify.js"
   },
   "bin": {
     "acorn": "./bin/acorn"
@@ -36,6 +38,7 @@
     "babelify": "^6.1.2",
     "browserify": "^10.2.4",
     "browserify-derequire": "^0.9.4",
+    "is-browser": "^2.0.1",
     "unicode-7.0.0": "~0.1.5"
   }
 }
