+++ test/browserify.html
@@ -1,0 +1,9 @@
+<!doctype html>
+<head>
+  <meta charset="utf-8">
+  <title>Acorn test suite</title>
+</head>
+<body>
+  <ul id="log"></ul>
+  <script src="browserify.js"></script>
+</body>
