+++ test/index.html
@@ -10,5 +10,6 @@
 </head>
 <body>
   <ul id="log"></ul>
-  <script src="run.js"></script>
+  <script src="runner.js"></script>
+  <script>run(acorn, window);</script>
 </body>
