+++ test/run.js
@@ -1,113 +1,13 @@
-(function() {
-  var driver, acorn;
+'use strict';
 
-  if (typeof require !== "undefined") {
-    driver = require("./driver.js");
-    require("./tests.js");
-    require("./tests-harmony.js");
-    require("babel-core/register")
-    acorn = require("../src")
-    require("../src/loose")
-  } else {
-    driver = window;
-    acorn = window.acorn;
-  }
+var runner = require('./runner.js');
 
-  var htmlLog = typeof document === "object" && document.getElementById('log');
-  var htmlGroup = htmlLog;
+var driver = require("./driver.js");
+require("./tests.js");
+require("./tests-harmony.js");
 
-  function group(name) {
-    if (htmlGroup) {
-      var parentGroup = htmlGroup;
-      htmlGroup = document.createElement("ul");
-      var item = document.createElement("li");
-      item.textContent = name;
-      item.appendChild(htmlGroup);
-      parentGroup.appendChild(item);
-    }
-    if (typeof console === "object" && console.group) {
-      console.group(name);
-    }
-  }
+require("babel-core/register");
+var acorn = require("../src");
+require("../src/loose");
 
-  function groupEnd() {
-    if (htmlGroup) {
-      htmlGroup = htmlGroup.parentElement.parentElement;
-    }
-    if (typeof console === "object" && console.groupEnd) {
-      console.groupEnd(name);
-    }
-  }
-
-  function log(title, message) {
-    if (htmlGroup) {
-      var elem = document.createElement("li");
-      elem.innerHTML = "<b>" + title + "</b> " + message;
-      htmlGroup.appendChild(elem);
-    }
-    if (typeof console === "object") console.log(title, message);
-  }
-
-  var stats, modes = {
-    Normal: {
-      config: {
-        parse: acorn.parse
-      }
-    },
-    Loose: {
-      config: {
-        parse: acorn.parse_dammit,
-        loose: true,
-        filter: function (test) {
-          var opts = test.options || {};
-          if (opts.loose === false) return false;
-          return (opts.ecmaVersion || 5) <= 6;
-        }
-      }
-    }
-  };
-
-  function report(state, code, message) {
-    if (state != "ok") {++stats.failed; log(code, message);}
-    ++stats.testsRun;
-  }
-
-  group("Errors");
-
-  for (var name in modes) {
-    group(name);
-    var mode = modes[name];
-    stats = mode.stats = {testsRun: 0, failed: 0};
-    var t0 = +new Date;
-    driver.runTests(mode.config, report);
-    mode.stats.duration = +new Date - t0;
-    groupEnd();
-  }
-
-  groupEnd();
-
-  function outputStats(name, stats) {
-    log(name + ":", stats.testsRun + " tests run in " + stats.duration + "ms; " +
-      (stats.failed ? stats.failed + " failures." : "all passed."));
-  }
-
-  var total = {testsRun: 0, failed: 0, duration: 0};
-
-  group("Stats");
-
-  for (var name in modes) {
-    var stats = modes[name].stats;
-    outputStats(name + " parser", stats);
-    for (var key in stats) total[key] += stats[key];
-  }
-
-  outputStats("Total", total);
-
-  groupEnd();
-
-  if (total.failed && typeof process === "object") {
-    process.stdout.write("", function() {
-      process.exit(1);
-    });
-  }
-})();
+runner(acorn, driver, 'Test Source');
